# crates/byos/src/infra/api/routes.rs
//! Route handlers for the public proposal API.

use {
    super::{
        AppState,
        dto::{
            CreateProposalRequest,
            CreateProposalResponse,
            GetProposalResponse,
            ListProposalsResponse,
            ProposalMetadata,
            parse_hex,
            parse_u256,
        },
        error::{Error, Kind},
    },
    crate::domain::proposal::{OrderUid, ProposalId, ProposalStatus},
    alloy::primitives::{Bytes, Signature, U256, keccak256},
    axum::{
        Json,
        extract::{Path, State},
        http::{HeaderMap, StatusCode},
        response::IntoResponse,
    },
    byos_common::{
        contracts::{Interaction, Proposal},
        eip712,
    },
    std::time::{SystemTime, UNIX_EPOCH},
};

// ---------------------------------------------------------------------------
// GET /healthz
// ---------------------------------------------------------------------------

pub async fn healthz() -> StatusCode {
    StatusCode::OK
}

// ---------------------------------------------------------------------------
// POST /proposals
// ---------------------------------------------------------------------------

pub async fn create_proposal(
    State(state): State<AppState>,
    Json(body): Json<CreateProposalRequest>,
) -> Result<impl IntoResponse, Error> {
    // 1. Parse and validate fields.
    let order_uid = OrderUid::from_hex(&body.order_uid)
        .map_err(|e| Error::new(Kind::BadRequest, format!("invalid orderUid: {e}")))?;

    let sell_amount = parse_u256(&body.sell_amount)
        .map_err(|_| Error::new(Kind::BadRequest, "invalid sellAmount"))?;
    let buy_amount = parse_u256(&body.buy_amount)
        .map_err(|_| Error::new(Kind::BadRequest, "invalid buyAmount"))?;
    let valid_until = parse_u256(&body.valid_until)
        .map_err(|_| Error::new(Kind::BadRequest, "invalid validUntil"))?;
    let nonce =
        parse_u256(&body.nonce).map_err(|_| Error::new(Kind::BadRequest, "invalid nonce"))?;

    let signature_bytes = parse_hex(&body.signature)
        .map_err(|_| Error::new(Kind::BadRequest, "invalid signature hex"))?;
    let signature = Signature::try_from(signature_bytes.as_slice())
        .map_err(|_| Error::from(Kind::InvalidSignature))?;

    // 2. Convert interactions.
    let interactions: Vec<Interaction> = body
        .interactions
        .iter()
        .map(Interaction::try_from)
        .collect::<Result<_, _>>()?;

    // 3. Compute hashes.
    let order_uid_hash = keccak256(order_uid.0);
    let interactions_hash = eip712::compute_interactions_hash(&interactions);

    // 4. Build the on-chain Proposal struct for signature recovery.
    let proposal = Proposal {
        orderUidHash: order_uid_hash,
        sellAmount: sell_amount,
        buyAmount: buy_amount,
        validUntil: valid_until,
        nonce,
    };

    // 5. Recover the sub-solver address.
    let sub_solver =
        eip712::recover_proposer(&signature, state.domain(), &proposal, interactions_hash)
            .map_err(|_| Error::from(Kind::SignatureRecoveryFailed))?;

    tracing::info!(%sub_solver, "proposal signature verified");

    // 6. Reject proposals that are already expired — no point accepting,
    // storing, and auditing a DOA proposal (ADR-0001) — or that outlive the
    // lifetime cap, which bounds simulation cost per proposal (ADR-0013).
    let now = SystemTime::now()
        .duration_since(UNIX_EPOCH)
        .expect("system clock before Unix epoch")
        .as_secs();
    if valid_until < U256::from(now) {
        return Err(Error::from(Kind::ProposalExpired));
    }
    if valid_until > U256::from(now.saturating_add(state.max_proposal_lifetime_secs())) {
        return Err(Error::from(Kind::ProposalLifetimeExceeded));
    }

    // 7. Store as Submitted. The background validator picks it up and flips
    // it to Active or Rejected; sub-solvers poll GET /proposal/{id} for the
    // verdict.
    let stored = crate::domain::proposal::Proposal {
        id: ProposalId(0),
        sub_solver,
        order_uid,
        order_uid_hash,
        sell_amount,
        buy_amount,
        // Token addresses come from the orderbook order during validation
        // (ADR-0012); placeholders until the first Accept verdict.
        sell_token: alloy::primitives::Address::ZERO,
        buy_token: alloy::primitives::Address::ZERO,
        interactions,
        interactions_hash,
        valid_until,
        nonce,
        signature: Bytes::from(signature_bytes),
        status: ProposalStatus::Submitted,
        rejection_reason: None,
        gas_used: None,
        trampoline: None,
        settlement_tx_hash: None,
    };

    let id = state.store().insert(stored).await.map_err(internal)?;

    tracing::info!(%id, %sub_solver, "proposal accepted for validation");

    Ok((StatusCode::ACCEPTED, Json(CreateProposalResponse { id })))
}

// ---------------------------------------------------------------------------
// GET /proposal/{id}
// ---------------------------------------------------------------------------

pub async fn get_proposal(
    State(state): State<AppState>,
    Path(id): Path<ProposalId>,
    headers: HeaderMap,
) -> Result<Json<GetProposalResponse>, Error> {
    let reader = authenticate_reader(&headers, state.domain())?;

    let proposal = state
        .store()
        .get(id)
        .await
        .map_err(internal)?
        // Non-owners get the same 404 as a genuine miss so proposal IDs
        // cannot be probed for existence (ADR-0011).
        .filter(|p| p.sub_solver == reader)
        .ok_or(Error::from(Kind::ProposalNotFound))?;

    Ok(Json(GetProposalResponse {
        id: proposal.id,
        sub_solver: proposal.sub_solver,
        order_uid: proposal.order_uid.to_string(),
        sell_amount: proposal.sell_amount.to_string(),
        buy_amount: proposal.buy_amount.to_string(),
        valid_until: proposal.valid_until.to_string(),
        status: proposal.status.to_string(),
        rejection_reason: proposal.rejection_reason,
        settlement_tx_hash: proposal.settlement_tx_hash.map(|t| format!("{t:#x}")),
    }))
}

// ---------------------------------------------------------------------------
// GET /proposals/{order_uid}
// ---------------------------------------------------------------------------

pub async fn list_proposals(
    State(state): State<AppState>,
    Path(order_uid_hex): Path<String>,
    headers: HeaderMap,
) -> Result<Json<ListProposalsResponse>, Error> {
    let reader = authenticate_reader(&headers, state.domain())?;

    let order_uid = OrderUid::from_hex(&order_uid_hex)
        .map_err(|e| Error::new(Kind::BadRequest, format!("invalid orderUid: {e}")))?;

    // Owner-scoped reads (ADR-0011): the store query only returns the
    // caller's own proposals on this order.
    let proposals = state
        .store()
        .list_by_order_uid_for_owner(&order_uid, reader)
        .await
        .map_err(internal)?;

    Ok(Json(ListProposalsResponse {
        proposals: proposals.iter().map(ProposalMetadata::from).collect(),
    }))
}

// ---------------------------------------------------------------------------
// GET /proposals/by-solver
// ---------------------------------------------------------------------------

/// Lists the caller's own active proposals. The caller's identity comes
/// entirely from the `X-Signature` header — there is no address parameter
/// (ADR-0011).
pub async fn list_proposals_by_solver(
    State(state): State<AppState>,
    headers: HeaderMap,
) -> Result<Json<ListProposalsResponse>, Error> {
    let reader = authenticate_reader(&headers, state.domain())?;

    let proposals = state
        .store()
        .list_by_sub_solver(reader)
        .await
        .map_err(internal)?;

    Ok(Json(ListProposalsResponse {
        proposals: proposals.iter().map(ProposalMetadata::from).collect(),
    }))
}

// ---------------------------------------------------------------------------
// DELETE /proposal/{id}
// ---------------------------------------------------------------------------

pub async fn cancel_proposal(
    State(state): State<AppState>,
    Path(id): Path<ProposalId>,
    headers: HeaderMap,
) -> Result<StatusCode, Error> {
    // 1. Extract signature from X-Signature header.
    let signature = signature_from_header(&headers)?;

    // 2. Recover signer from CancelProposal EIP-712 message.
    let signer = eip712::recover_canceller(&signature, state.domain(), U256::from(id.0))
        .map_err(|_| Error::from(Kind::SignatureRecoveryFailed))?;

    // 3. Cancel the proposal (store checks ownership). Non-owners get the
    // same 404 as a genuine miss — a 403 would be an existence oracle for
    // proposal ids, the thing ADR-0011's owner-scoped 404 exists to prevent.
    use crate::infra::storage::StoreError;
    state
        .store()
        .cancel(id, signer)
        .await
        .map_err(|e| match e {
            StoreError::NotFound(_) | StoreError::NotOwner(_, _) => {
                Error::from(Kind::ProposalNotFound)
            }
            StoreError::StaleTransition { .. } => Error::from(Kind::ProposalNotCancellable),
            e => internal(e),
        })?;

    tracing::info!(%id, %signer, "proposal cancelled");

    Ok(StatusCode::NO_CONTENT)
}

// ---------------------------------------------------------------------------
// Helpers
// ---------------------------------------------------------------------------

/// A store failure the caller can do nothing about: log it, answer 500.
fn internal(e: crate::infra::storage::StoreError) -> Error {
    tracing::error!(%e, "proposal store error");
    Error::from(Kind::Internal)
}

/// Authenticates a GET request: extracts the `X-Signature` bearer token and
/// recovers the reader's address from the `ReadAuth` EIP-712 message
/// (ADR-0011). The returned address scopes what the caller may read.
fn authenticate_reader(
    headers: &HeaderMap,
    domain: &alloy::sol_types::Eip712Domain,
) -> Result<alloy::primitives::Address, Error> {
    let signature = signature_from_header(headers)?;
    eip712::recover_reader(&signature, domain)
        .map_err(|_| Error::from(Kind::SignatureRecoveryFailed))
}

fn signature_from_header(headers: &HeaderMap) -> Result<Signature, Error> {
    let value = headers.get("X-Signature").ok_or(Error::new(
        Kind::InvalidSignature,
        "missing X-Signature header",
    ))?;
    let hex_str = value
        .to_str()
        .map_err(|_| Error::new(Kind::InvalidSignature, "X-Signature is not valid UTF-8"))?;
    let bytes = parse_hex(hex_str)
        .map_err(|_| Error::new(Kind::InvalidSignature, "X-Signature is not valid hex"))?;
    Signature::try_from(bytes.as_slice())
        .map_err(|_| Error::new(Kind::InvalidSignature, "invalid signature bytes"))
}